+++ packages/a11y-base/README.md
@@ -1,0 +1,14 @@
+# @vaadin/a11y-base
+
+A set of accessibility helpers, mixins and controllers used by Vaadin components.
+
+## Contributing
+
+Read the [contributing guide](https://vaadin.com/docs/latest/contributing/overview) to learn about our development process, how to propose bugfixes and improvements, and how to test your changes to Vaadin components.
+
+## License
+
+Apache License 2.0
+
+Vaadin collects usage statistics at development time to improve this product.
+For details and to opt-out, see https://github.com/vaadin/vaadin-usage-statistics.
