+++ packages/a11y-base/index.js
@@ -1,0 +1,1 @@
+export { announce } from './src/announce.js';
