+++ packages/a11y-base/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@vaadin/a11y-base",
+  "version": "24.0.0-beta1",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "Vaadin accessibility helpers, mixins and controllers",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/vaadin/web-components.git",
+    "directory": "packages/a11y-base"
+  },
+  "author": "Vaadin Ltd",
+  "homepage": "https://vaadin.com/components",
+  "bugs": {
+    "url": "https://github.com/vaadin/web-components"
+  },
+  "main": "index.js",
+  "module": "index.js",
+  "type": "module",
+  "files": [
+    "index.d.ts",
+    "index.js",
+    "src"
+  ],
+  "keywords": [
+    "Vaadin",
+    "web-components",
+    "web-component"
+  ],
+  "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
+    "@polymer/polymer": "^3.0.0",
+    "@vaadin/component-base": "24.0.0-beta1",
+    "lit": "^2.0.0"
+  },
+  "devDependencies": {
+    "@esm-bundle/chai": "^4.3.4",
+    "@vaadin/testing-helpers": "^0.4.0",
+    "sinon": "^13.0.2"
+  }
+}
