+++ packages/a11y-base/src/announce.d.ts
